backend/requirements.txt
django==4.1.5
psycopg2==2.9.5
djangorestframework==3.14.0
requests==2.28.1
django-cors-headers==3.13.0
selenium==4.7.2
bs4==0.0.1
channels[daphne]==4.0.0
channels_redis==4.0.0
djangorestframework-simplejwt==5.2.2